First iteration of script to migrate redmine issues and milestones:
This script migrates all issues against a Redmine Project to a GitLab
Project. Issues cannot be created against a group, that could prove
to be a problem for some projects.

Redmine Project Versions are mapped to GitLab Project Milestones, again,
this could prove a problem for some services that use group versions, not
tested against them so not entirely sure what is possible.