Create Milestones against the parent group:
We migrate Project issues, but Milestones they belong to, need to be
at the Group level. This is because there could be many projects that
combine to make a service, so Milestones need to be able to span
multiple Projects.

This causes a new check to be made, we shouldn't try to create an
already existing Milestone. I think there is a bug if there are
Redmine Versions, but they have already got Gitlab Milestones, not
sure if issues will then be created, needs more testing.